requirements.txt
certifi==2018.1.18
chardet==3.0.4
clearbit==0.1.7
click==6.7
Django==2.0.3
django-crispy-forms==1.7.2
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
Faker==0.8.12
idna==2.6
pyhunter==0.3
PyJWT==1.6.1
python-dateutil==2.7.2
pytz==2018.3
requests==2.18.4
six==1.11.0
termcolor==1.1.0
text-unidecode==1.2
urllib3==1.22
